cache directory related to paperclip location
added system property `paperclip.enterprise`, when true vanilla jar and
patched jar will be stored in `cache/` directory related to where the
paperclip jar file is located.

this is to keep configuration separated from actual server files. makes
it easier to build a docker container to run the server in.